Controller: fix 'duplicate working' when validating
If validating multiple package at the same time, an error
was visible when starting VM. Only the first package
was really validated (and possibly published), and the
others were all failing.

Simply manage the working repo outside of the loop
to fix the issue.

This loop should be cleaned and improved in the future.
More stuff should be prepared outside of the loop and
publication only done in one step, after the loop.

Change-Id: Ibaaa2b89103f24cdb818e666fa53e5ab3d9cafe9