Controller: fix bug when testing already installed packages
Before this patch, 'yum history undo' was launched every time.
But when a package was already up to date on the system, it may lead to
reinstall/uninstall last packages depending on what was done previously.

With this patch, 'yum history undo' is done only if a transaction was
made during the test phase.

Change-Id: I1684b33cd4bbd284bdfacd56c7e33f181ff926fd