RPM: GPG signature support
This commit adds an optional dict configuration parameter to specify GPG
settings. This dict accepts 3 keys to specify keyring, GPG key ID and an
optional passphrase used to encrypt this key. When these parameters are
defined and when -s, --sign option is set on build, validate and
validdiff actions command line, Rift uses the corresponding GPG key to
sign built source and binary RPM packages. When the keyring and key ID
is defined without passphrase, the key is encrypted and the passphrase
is not present in agent cache, GPG agent asks user for the passphrase
interactively. Note that -s, --sign option automatically implies -p,
--publish.

New rift sign action is also introduced to sign arbitrary source or
binary RPM package.

An additional is_signed boolean attribute is added to RPM objects to
indicate whether they are signed with a key or not.

Unit tests are added to test all methods of RPM class, including the new
sign() method in all its code paths with or without GPG agent. In order
to facilitate definition and execution of these tests, very small source
and binary RPM packages are added in tests/materials/ folder. Unit tests
are also added to test possible and refused combinations of GPG
settings.

Example project configuration file is also updated with example gpg
settings hash.

Change-Id: I46a23587de5dc218489d0d4bd11b2e303fb65bc2