Controller: validdiff removes pkgs from repo
With this commit, in addition to updated packages, validdiff searches
for removed packages in patches. A package is considered removed as soon
as its info.yaml metadata file is removed by the patch.

In such case, if --publish argument is enabled and a working repository
is declared in configuration, validdiff removes RPM packages generated
by this source package from this working repository. Source package and
all binary packages generated by this source package are removed.

Some unit tests are modified and some tests are added to validate the
new behaviour.

Change-Id: I0d71e8a2407d32d4e864ea95c8142d0c046c8df0