Controller: Display an error if the VM is already running
If the VM is already running and the rift validate is utilized,
then rift will display that the VM is already running. However it
will also indicate that all packages wered checked successfully.

By checking the return code of the method action_test(), the user is
warned that the tests did not sucess.

Tests should always be launched in a 'clean' VM to be sure that the package
integrates correctly into Ocean packages.

Change-Id: I9e95d2aadd8543e44552e3894b592a5f3883aaa1