Config: Add VM memory option
Currently the memory of the rift VM is always 8192 mb because it's
hardcoded in VM.py (Line 140).

This not a good idea because partners may not have this amount of
memory or in the future packages may require more memory to test
them.

This commit introduce the 'vm_memory' configuration option to define
how much memory the rift VM can use. This option is directly passed
to qemu (-m option)

Change-Id: I25288aa62b98de81e03da837dba8527251b02ca4