Controller: refactor actions functions
Introduce action_build(), action_test() and action_validate() functions
in order to simplify and make more consistent bug action() function.

Another new function staff_modules() is also introduced to get Staff and
Modules objects when required.

Some functions are renamed to avoid action_*() function to be called for
other actions (eg. action_build() called by action_validate()) and
reserve action functions to the corresponding CLI actions.

This refactoring is performed in preparation of introduction of
multi-architectures support, that is the reason why it is restricted to
controller actions where multi-architectures support is expected.

Change-Id: Ifb17064b4b0a520176004d6aae9641b1fe5c92ba