Add multi-architectures support
The type of project configuration arch parameter is changed from a
string to a list of string to give the possibility to specify several
supported CPU architectures for a project.

Rift actions such as build, test, validate and validdiff now iterate
over the list of all supported architecture to perform the action for
all of them.

The vm action has a new -a,--arch argument to specify the architecture.
This argument is required when multiple supported architectures are set
in project configuration. Otherwise, the argument is optional, the value
is the only supported architecture by default.

Config class is updated to check list type for arch and support multiple
architectures in architecture specific parameters definitions.

VM, Mock, TestResults and TestCase classes are enriched with an
architecture argument so they can be instanciated for multiple
architectures.

Some unit tests to run build, test and validate actions are updated to
run on multiple architectures and verify proper iteration over the list
of architectures. Tests on Config and ProjectArchRepositories are also
updated to validate behavior with multiple architectures. Unit tests on
vm action are added to check --arch option and Mock tests are updated
with new architecture argument.

Some tests require qemu-user-static for mock to run. They are
automatically skipped if not installed.

Change-Id: I584e8d222b1688fd5bf1a27e1bb93d124487feb2