VM: add build command
Add rift vm build command to build the virtual machine image designed
for tests purpose. The image is built based on distribution cloud image
available on the internet. The image is customized with cloud-init and
a build-post script stored in project folder. Generic working examples
of these files are added in template/ subfolder.

Users must give the URL of the cloud image in argument of the command.
Rift maintains a cache of the downloaded image. The download can be
forced with --force option.

The user must precise the path of the generated image with either
-o,--output or --deploy option. With the first option, the user can
give any arbitrary path. With the second option, the project vm_image
is generated. If the targeted path already exists, the user is asked
interactively whether the file must be overwritten or not.

Tests are updated accordingly to fulfill new configuration parameters
requirements, validate all command line parameters options, validate
some VM build cases and error management. Some tests require qemu-img
executable, they are automatically skipped if the executable is not
found.

Change-Id: Id900eb34ba02a23bbe5ae81adea7f6e495598d62