Repository: refactor classes for multiarch
Classes in Repository module are renamed and refactored to better
support multiple architectures.

Class RemoteRepository is renamed ConsumableRepository, it handles
one repository (specific to an architecture) which can be consumed by
yum/dnf. Class Repository is renamed LocalRepository, it handles a local
repository which can be populated with RPM packages. The inheritance
between Repository and RemoteRepository is removed, LocalRepository is
now a set to ConsumableRepository (one per architecture).

VM and Mock classes do not loosely and opportunistically create
(local) repositories anymore, they check if the local repositories
actually exist and raise error if not.

Working repository creation is now the responsability of
ProjectArchRepositories class.

Tests are updated to cover all new refactored repository classes and
cover new error cases in VM and Mock classes. Tiny source and binary RPM
packages are added in tests/materials/ folder for the purpose of the
tests.

Change-Id: Ic9ac18b4073a0fe8e35d3d81600d28cf3ece9562