Config: merge dict from multiple files
When a configuration parameter in dictionnary format is defined in
multiple configuration files (eg. project.conf and local.conf), its
resulting value is the merge of the values from all files.

This is typically useful for repositories as it makes possible to add a
single repository in local.conf without requiring duplication of all
repositories from project.conf.

Unit tests are added to validate the correct behavior of this feature.

Change-Id: I9e6ce847216bf0635c38b9a7a232d9cf5f0f8648