VM: allow one VM per user/arch/version
The VM ID is now a hash of the user/arch/version combination. The
default port is allocated in a range using a modulo of this hash
value. Thus, two virtual machines with the same combination have the
same hash and default port. If either user, arch or version is
different, the hash and the default port changes.

In configuration, new vm_port_range dict parameter is introduced to
define min/max limits of the allocated range of TCP ports. In the same
time, the vm_port parameter is removed from configuration as there is no
actual use case for this parameter anymore.

Unit test are added to validate these design rules.

Change-Id: I7b915f1647067e8faadf11b39fd07c9afd1852bc