Controller: rename parse_options to make_parser
To use argparse-manpage, we need a function returning the parser object
of rift.
parse_options was almost ready to do that, we rename it and make the
parsing outside the function to make it compatible with
argparse-manpage.

see: https://gitlab.ccc.ocre.cea.fr/cedeyna/test-ocean/-/issues/166
for more details.

Change-Id: Ib9c099542ae0044fda66339593ca054133edd496