sync: introduce repository synchronization feature
Add possibility to mirror remote packages repository on local
filesystem.

The feature is managed by new sync module. The remote repository to
synchronize must be declared in project configuration. Three
synchronization methods are supported with this initial implementation:

- lftp: Execute lftp command to mirror a subdirectory recursively.
- epel: Designed to support Fedora EPEL repository layout, it downloads
  all files referenced in an index under a given subdirectory.
- dnf (default): Download dnf repository metadata to discover all RPM
  packages.

With all methods, it is possible to specify inclusion and exclusion
patterns to finely control which files are actually synchronized.

It is possible to specify a list of repositories to synchronize in
command line argument. By default, all repositories defined with
synchronization configuration parameters are synchronized.

The utils module is introduced to avoid code duplication with VM module.

Unit tests are added to cover new code introduced for this feature.
Tiny source and binary RPM packages are added in tests/materials/ folder
for the purpose of the tests.

Change-Id: Id49f80d20b9c6d1bc3ae31ecf377b42085b261c2