run: introduce run module
This commit introduces new run module with run_command() function that
is able to both capture output of a subprocess and stream its output in
current process stdout/stderr. The function returns a RunResult named
tuple with the returncode, out and err streams as string (or None when
capture is disabled).

This run_command() function is used in VM.cmd(), VM.run_test() and
Mock._exec() as a high-level replacement for raw subprocess.Popen()
calls.

Due to the fact that VM.cmd() and VM.run_test() now return a RunResult
instead of return code integer, some changes had to be done in
Controller module.

Some unit tests are added to validate the new function with all
combinations of arguments.

Change-Id: I032dd992de378a5300c255949f6240631f0da5a3