Controller: RPM spec parse error in test results
Detect RPM spec file parse error in build/validate/test actions,
log an error and report the error in tests results. This notably makes
the error reported in junit files eventually.

Note that for the test action, a dummy parse test case in introduced
specifically to report this error, only when it fails. For build and
validate actions, the error is reported in build test case.

Change-Id: Ibeeaa3fbeefb1e5e0bb76bbe77d28de0b82858da