Config: add possibility to deprecate parameters
This is now possible to mark configuration parameters as deprecated, by
specifying the name of their replacement. Arbitrary levels of dict keys
can be specified in the name using dot separator (ex: hash1.hash2.key3).

A FutureWarning is emitted when deprecated parameter is loaded, to alert
user of the change and inform of the new parameter name.

When both the deprecated and replacement parameters are defined, the
deprecated parameter is ignored.

Unit tests are introduced to cover all cases of this feature.