template/build-post.sh: effectively disable default packages
Initialy, the post-script have to disable default repositories of the
distribution to only use the one provided by rift.
The swith to almalinux broke this script.

I added the Rocky add Almalinux repository detection to fix that.

If there is no more repository yum update will fail and this is the
expected behaviour, so we don't return an error in that case.

Change-Id: Iec399f43a928321926751068c2b0126d68ba5af4