Config: check all required params recursively
Instead of checking for required in params in sub dictionnaries at load
time, check them recursively after everything is loaded.

This change notably fixes bug in case required parameter in a sub
dictionnary is not defined in a configuration file but defined in
another file and in merged options eventually.