RPM: workaround timezone bug in spec parsing
The RPM library change the timezone when parsing spec file but old
versions of this library (<4.19) miss to restore the timezone properly.
This has the side effect to change the timezone for the end of Rift
execution, which could lead to weird behavior when parsing or
manipulating dates and times. In order to workaround this, Rift now
saves and restores the timezone itself, just to make sure everything is
in order after parsing the spec file.