Tests: disable mock bootstrap image
Mock bootstrap image feature leads to some unexplained behavior when
running multi-arch builds with podman in docker in GitHub actions. As an
example, builds of foreign architecture fail with file:// repository
metadata not found during bootstrap when local repository is specified
in build environment, despite a valid multi-arch local repository and
correct usage of $basearch in repository path to handle architecture
difference between bootstrap image and build environment.

This might be related to this issue:
https://github.com/rpm-software-management/mock/issues/1570

Bootstrap feature is not required by tests as testing environment
support all expected rpm features so the quick and easy fix is to
disable this feature in tests.