RPM: Fix changelog --bump option (#13)
To compute the bumped release version of a Specfile, the rstrip method was used to remove the dist tag before computing the bump (We don't want to bump the dist tag, el9 to el10 for instance). However, the rstrip version remove a set of character in the totality of the string instead of only the suffix. The removesuffix method is more suitable.

Further documentation : https://docs.python.org/3/library/stdtypes.html#str.rstrip