Tests: actual source tarball for testing package
Instead of creating a flat file with .tar.gz extension, create an actual
gzip compressed tarball with top folder and dummy source file. This is
useful for tests that expect such tarball.

The name of the tarball top folder can be changed with src_top_dir
argument. By default, the name of this top folder is $name-$version.