Package: move local test logic out of VM class
VM.run_test() was responsible of running local tests, designed to be
executed on host. The responsability has been transfered to
ActionableArchPackage class because this need is shared by all supported
packages formats. The VM class is now only responsible to provide the
dict of shell functions with utilities to handle the VM in the tests.