Annex.py: Update hash algorithm and switch to UNIX timestamps (#27)
* Annex.py: Update hash algorithm and switch to UNIX timestamps

Switch from MD5 to SHA3-256 to avoid any risk of collision. MD5 is still supported for retrieving from the annex, pushing is only using SHA3-256
Switch to %c date format to a UNIX timestamp to allow collaboration with people with different date format / timezome. The old format is supported for reading, the push only UNIX timestamp
---------

Co-authored-by: VALERI Yoann <yoann.valeri@cea.fr>