vm: increase grace period for vm creation
Increase the grace period during which we check the VM to build has
started. Currently, for x86_64 architectures, it waits up to 25 seconds
(or 100 seconds for aarch64). With this change, the grace period is
increased to 50 and 200 seconds by default. A parameter is planned for
the feature to make this more dynamic.

Change-Id: I56d630165fa7f36f59bdcc3224d98033a2e6df7b