VM: support remote VM image
The vm image configuration parameter can now be a http(s) URL to a
remote image, as an alternative to a path to a local image on
filesystem. In this case, rift downloads a copy of the image in
temporary directory before starting it.

The path to the local copy is deterministic, in order to avoid
redownloading the image on every rift successive runs.

When a remote image URL is set in configuration, rift vm build --deploy
cannot be used. Users are invited to use -o, --output option instead.